+++ packages/snap/snap.manifest.json
@@ -7,7 +7,7 @@
     "url": "https://github.com/MetaMask/template-snap-monorepo.git"
   },
   "source": {
-    "shasum": "GQqk/ohMeoiAtj9GK6kYvzf4RcLXSqfohrV9FKyZIVY=",
+    "shasum": "2URbO7qbOW+kOm0rY5bjnXLvLcdcHJNJ8B/UyPauvfA=",
     "location": {
       "npm": {
         "filePath": "dist/bundle.js",
