+++ packages/snap/src/helpers/store-managment.helper.ts
@@ -1,0 +1,11 @@
+export async function getData() {
+  try {
+    const data = await snap.request({
+      method: 'snap_manageState',
+      params: { operation: 'get' },
+    });
+    return data;
+  } catch (error) {
+    return undefined;
+  }
+}
