+++ packages/snap/src/types/text-and-copy.type.ts
@@ -1,0 +1,6 @@
+import { NodeType } from '@metamask/snaps-ui';
+
+export type arraySharesToTextAndCopyType = {
+  value: string;
+  type: NodeType.Copyable | NodeType.Text;
+};
